+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0 - 2023-11-07
+
+* Update Example
+* Update README.md to include more information
+* New hasRole method added to `PermissionPolicy` class
+* Initialize Nylo in the `addRoles` method
+
 ## 0.1.0 - 2023-11-04
 
 * Initial release